% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterSSH.R
\name{initCluster}
\alias{initCluster}
\title{Initialize a Cluster}
\usage{
initCluster(userFuns, remFun = NULL, sshOpts = NULL)
}
\arguments{
\item{userFuns}{a list which contain at least the functions \code{printId}, \code{start}, and \code{isRunning}}

\item{remFun}{object as returned from the \code{initSSH} function of the \code{remoteFunctionSSH} package}

\item{sshOpts}{named list of arguments passed to \code{remoteFunctionSSH}}
}
\value{
Returns a list which contains the following functions:

\code{start(fun, input, inputsPerTask=1, runOpts)}

\code{eval(fun, input, inputsPerTask=1, runOpts, pollTime)}

\code{result(job, selection=TRUE)}

\code{numberOfTasks(job)}

\code{closeCon()}

\code{status(job, raw=FALSE)}

The meaning of the arguments is
\tabular{ll}{
  \code{fun} \tab a function taking as argument a list \code{input};
  it must return a list of the same length as \code{input}
  where each element of the resulting list contains
  the result associated with the respective element in \code{input}\cr
  \code{input} \tab a list where each element is an input for a calculation\cr
  \code{inputsPerTask} \tab how many elements in \code{input} should be submitted as one task\cr
  \code{runOpts} \tab a list with arguments for the SGE engine, which are
  \code{cputime} (seconds), \code{vmem}, and \code{fsize} (megabytes). \cr
  \code{pollTime} \tab pause in seconds between two consecutive queries of the job status\cr
  \code{selection} \tab select for which tasks the results should returned\cr
  \code{raw} \tab should raw output of qstat be returned
}

The return values of the functions are
\tabular{ll}{
\code{start} \tab returns \code{job} object\cr
\code{eval} \tab returns the results associated with the elements of \code{input}\cr
\code{result} \tab returns the same as \code{eval}\cr
\code{status} \tab returns either a dataframe with information about the tasks or
 a character vector with the raw output of \code{qstat}
}
}
\description{
Initialize a cluster running a job scheduling engine such as the Sun Grid Engine (SGE) software
}
